AI-ContentGen-Pro/requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
flask>=3.0.0
requests>=2.31.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
prettytable>=3.9.0
tiktoken>=0.5.0
